Updated file 7

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -9,7 +9,7 @@
  */
 int print_last_digit(int n)
 {
-	int last_digit;
+	long int last_digit;
 
 	if (n < 0)
 	{
